Created branch. Added .gitignore Configured release build. Moved source files. Fixed name order for girls. Added girl name instead "P1" and "P2" in output console.
