+++ .gitignore
@@ -1,0 +1,3 @@
+bin
+build
+.vs
