--- LabTool/Debug/LabTool.log
@@ -1,2 +1,0 @@
-﻿  functions.cpp
-  LabTool.vcxproj -> C:\Users\Paul\source\repos\LabTool\Debug\LabTool.dll
