--- LabTool/Debug/vcpkg.applocal.log
@@ -1,1 +1,0 @@
-
