+++ LabTool/src/address.h
@@ -1,0 +1,333 @@
+﻿#pragma once
+#ifndef address_h__
+#define address_h__
+
+//#define WINDOW_CLASS            g_ini.GetValue("WINDOW_CLASS")
+//#define WINDOW_TEXT             g_ini.GetValue("WINDOW_TEXT")
+#define ADDR_PBATTLEMGR         addr[address::PBATTLEMGR]
+//#define ADDR_PNETOBJECT         addr[address::PNETOBJECT]
+#define ADDR_COMMMODE           addr[address::COMMMODE]
+//#define ADDR_LCHARID            addr[address::LCHARID]
+//#define ADDR_RCHARID            addr[address::RCHARID]
+#define ADDR_SCENEID            addr[address::SCENEID]
+#define ADDR_LCHAROFS           addr[address::LCHAROFS]
+#define ADDR_RCHAROFS           addr[address::RCHAROFS]
+#define ADDR_BTLMODEOFS         addr[address::BTLMODEOFS]
+#define ADDR_LPROFOFS           addr[address::LPROFOFS]
+#define ADDR_RPROFOFS           addr[address::RPROFOFS]
+#define ADDR_NETUDPOFS          addr[address::NETUDPOFS]
+#define ADDR_ADRBEGOFS          addr[address::ADRBEGOFS]
+#define ADDR_TOADDROFS          addr[address::TOADDROFS]
+#define ADDR_WINCNTOFS          addr[address::WINCNTOFS]
+
+#define ADDR_POINTXOFS          addr[address::POINTXOFS]
+#define ADDR_POINTYOFS          addr[address::POINTYOFS]
+#define ADDR_HPOFS              addr[address::HPOFS]
+#define ADDR_TIMECOUNT          addr[address::TIMECOUNT]
+#define ADDR_SPEEDXOFS          addr[address::SPEEDXOFS]
+#define ADDR_SPEEDYOFS          addr[address::SPEEDYOFS]
+#define ADDR_ACTIONIDOFS        addr[address::ACTIONIDOFS]
+#define ADDR_FRAMECOUNTOFS      addr[address::FRAMECOUNTOFS]
+#define ADDR_WEATHER            addr[address::WEATHER]
+#define ADDR_DISPLAYWEATHER     addr[address::DISPLAYWEATHER]
+#define ADDR_WEATHERCOUNTER     addr[address::WEATHERCOUNTER]
+#define ADDR_ACTIVEFLAG         addr[address::ACTIVEFLAG]
+#define ADDR_1PKEYMAP           addr[address::KEYMAP1P]
+#define ADDR_2PKEYMAP           addr[address::KEYMAP2P]
+#define ADDR_GAMEDATAMGR        addr[address::GAMEDATAMGR]
+#define ADDR_CHAROBJHEADOFS     addr[address::CHAROBJHEADOFS]
+#define ADDR_CHAROBJTAILOFS     addr[address::CHAROBJTAILOFS]
+#define ADDR_OBJLISTMGR         addr[address::OBJLISTMGR]
+#define ADDR_OBJPROJOFS         addr[address::OBJPROJOFS]
+#define ADDR_ATTACKAREAOFS      addr[address::ATTACKAREAOFS]
+#define ADDR_ATTACKAREA2OFS     addr[address::ATTACKAREA2OFS]
+#define ADDR_ATTACKAREACOUNTOFS addr[address::ATTACKAREACOUNTOFS]
+#define ADDR_HITAREACOUNTOFS    addr[address::HITAREACOUNTOFS]
+#define ADDR_DIRECTIONOFS       addr[address::DIRECTIONOFS]
+#define ADDR_IMAGESTRUCTOFS     addr[address::IMAGESTRUCTOFS]
+#define ADDR_IMAGENUMBEROFS     addr[address::IMAGENUMBEROFS]
+#define ADDR_HITSTOPOFS         addr[address::HITSTOPOFS]
+#define ADDR_AIRDASHCOUNTOFS    addr[address::AIRDASHCOUNTOFS]
+#define ADDR_REIPOWEROFS        addr[address::REIPOWEROFS]
+#define ADDR_MAXREIPOWEROFS     addr[address::MAXREIPOWEROFS]
+#define ADDR_CORRECTIONOFS      addr[address::CORRECTIONOFS]
+#define ADDR_COMBORATEOFS       addr[address::COMBORATEOFS]
+#define ADDR_COMBOCOUNTOFS      addr[address::COMBOCOUNTOFS]
+#define ADDR_COMBODAMAGEOFS     addr[address::COMBODAMAGEOFS]
+#define ADDR_COMBOLIMITOFS      addr[address::COMBOLIMITOFS]
+#define ADDR_CARDGAUGEOFS       addr[address::CARDGAUGEOFS]
+#define ADDR_CARDCOUNTOFS       addr[address::CARDCOUNTOFS]
+#define ADDR_CARDCOUNT2OFS      addr[address::CARDCOUNT2OFS]
+#define ADDR_HANDCARDBASEOFS    addr[address::HANDCARDBASEOFS]
+#define ADDR_SELECTCARDOFS      addr[address::SELECTCARDOFS]
+#define ADDR_HANDCARDMAXOFS     addr[address::HANDCARDMAXOFS]
+#define ADDR_SKILLLEVELMAPOFS   addr[address::SKILLLEVELMAPOFS]
+#define ADDR_MOTIONSTRUCTOFS    addr[address::MOTIONSTRUCTOFS]
+#define ADDR_HITAREAOFS         addr[address::HITAREAOFS]
+#define ADDR_LPROFNAME          addr[address::LPROFNAME]
+#define ADDR_RPROFNAME          addr[address::RPROFNAME]
+#define ADDR_ATTACKPOWEROFS     addr[address::ATTACKPOWEROFS]
+#define ADDR_DEFENSEPOWEROFS    addr[address::DEFENSEPOWEROFS]
+#define ADDR_DEFENSEDOLLOFS     addr[address::DEFENSEDOLLOFS]
+#define ADDR_ATTACKSTICKOFS     addr[address::ATTACKSTICKOFS]
+#define ADDR_KOKUSHIOFS         addr[address::KOKUSHIOFS]
+#define ADDR_UTHIWAOFS          addr[address::UTHIWAOFS]
+#define ADDR_SPEEDPOWEROFS      addr[address::SPEEDPOWEROFS]
+#define ADDR_GRIMOREOFS         addr[address::GRIMOREOFS]
+#define ADDR_DROPWATEROFS       addr[address::DROPWATEROFS]
+#define ADDR_DROPWATERTIMEOFS   addr[address::DROPWATERTIMEOFS]
+#define ADDR_RYUUSEIOFS         addr[address::RYUUSEIOFS]
+#define ADDR_TIMESTOPOFS        addr[address::TIMESTOPOFS]
+#define ADDR_KANAKOOFS          addr[address::KANAKOOFS]
+#define ADDR_SUWAKOOFS          addr[address::SUWAKOOFS]
+#define ADDR_STAGENUMBER        addr[address::STAGENUMBER]
+#define ADDR_STAGENUMBER2       addr[address::STAGENUMBER2]
+#define ADDR_BGMNUMBER          addr[address::BGMNUMBER]
+#define ADDR_SAKUYAWORLDOFS     addr[address::SAKUYAWORLDOFS]
+#define ADDR_PRIVATESQOFS       addr[address::PRIVATESQOFS]
+#define ADDR_TENSEITIMEOFS      addr[address::TENSEITIMEOFS]
+#define ADDR_TENSEINUMOFS       addr[address::TENSEINUMOFS]
+#define ADDR_TENSEIONOFS        addr[address::TENSEIONOFS]
+#define ADDR_LIFERECOVERYOFS    addr[address::LIFERECOVERYOFS]
+#define ADDR_TEKETENOFS         addr[address::TEKETENOFS]
+#define ADDR_ORRERYOFS          addr[address::ORRERYOFS]
+#define ADDR_PHILOSOPHEROFS     addr[address::PHILOSOPHEROFS]
+#define ADDR_MUNENOFS           addr[address::MUNENOFS]
+#define ADDR_HIHISOUOFS         addr[address::HIHISOUOFS]
+#define ADDR_FIELDREDOFS        addr[address::FIELDREDOFS]
+#define ADDR_FIELDPURPLEOFS     addr[address::FIELDPURPLEOFS]
+#define ADDR_MPPOFS             addr[address::MPPOFS]
+#define ADDR_FIELDRED2OFS       addr[address::FIELDRED2OFS]
+#define ADDR_HITAREAFLAGOFS     addr[address::HITAREAFLAGOFS]
+#define ADDR_HITAREA2OFS        addr[address::HITAREA2OFS]
+#define ADDR_FRAMEFLAGSOFS      addr[address::FRAMEFLAGSOFS]
+#define ADDR_ATTACKFLAGSOFS     addr[address::ATTACKFLAGSOFS]
+#define ADDR_DIAHARDOFS         addr[address::DIAHARDOFS]
+#define ADDR_REISTOPROFS        addr[address::REISTOPROFS]
+#define ADDR_KEYMGROFS          addr[address::KEYMGROFS]
+#define ADDR_KEYMAPOFS          addr[address::KEYMAPOFS]
+#define ADDR_ACTIONBLOCKIDOFS   addr[address::ACTIONBLOCKIDOFS]
+
+class address /*: boost::noncopyable*/ {
+public:
+	enum {
+		PBATTLEMGR = 0, PNETOBJECT, COMMMODE,
+		LCHARID, RCHARID, SCENEID,
+		LCHAROFS, RCHAROFS, BTLMODEOFS,
+		LPROFOFS, RPROFOFS, NETUDPOFS,
+		ADRBEGOFS, TOADDROFS, WINCNTOFS,
+
+		POINTXOFS, POINTYOFS, HPOFS,
+		TIMECOUNT, SPEEDXOFS, SPEEDYOFS,
+		ACTIONIDOFS, FRAMECOUNTOFS, WEATHER,
+		DISPLAYWEATHER, WEATHERCOUNTER, ACTIVEFLAG,
+		KEYMAP1P, KEYMAP2P, GAMEDATAMGR,
+		CHAROBJHEADOFS, CHAROBJTAILOFS, OBJLISTMGR,
+		OBJPROJOFS, ATTACKAREAOFS, ATTACKAREA2OFS,
+		ATTACKAREACOUNTOFS, HITAREACOUNTOFS, DIRECTIONOFS,
+		IMAGESTRUCTOFS, IMAGENUMBEROFS, HITSTOPOFS,
+		AIRDASHCOUNTOFS, REIPOWEROFS, MAXREIPOWEROFS,
+		CORRECTIONOFS, COMBORATEOFS, COMBOCOUNTOFS,
+		COMBODAMAGEOFS, COMBOLIMITOFS, CARDGAUGEOFS,
+		CARDCOUNTOFS, CARDCOUNT2OFS, HANDCARDBASEOFS,
+		SELECTCARDOFS, HANDCARDMAXOFS, SKILLLEVELMAPOFS,
+		MOTIONSTRUCTOFS, HITAREAOFS, LPROFNAME,
+		RPROFNAME, ATTACKPOWEROFS, DEFENSEPOWEROFS,
+		DEFENSEDOLLOFS, ATTACKSTICKOFS, KOKUSHIOFS,
+		UTHIWAOFS, SPEEDPOWEROFS, GRIMOREOFS,
+		DROPWATEROFS, DROPWATERTIMEOFS, RYUUSEIOFS,
+		TIMESTOPOFS, KANAKOOFS, SUWAKOOFS,
+		STAGENUMBER, STAGENUMBER2, BGMNUMBER,
+		SAKUYAWORLDOFS, PRIVATESQOFS, TENSEITIMEOFS,
+		TENSEINUMOFS, TENSEIONOFS, LIFERECOVERYOFS,
+		TEKETENOFS, ORRERYOFS, PHILOSOPHEROFS,
+		MUNENOFS, HIHISOUOFS, FIELDREDOFS,
+		FIELDPURPLEOFS, MPPOFS, FIELDRED2OFS,
+		HITAREAFLAGOFS, HITAREA2OFS, FRAMEFLAGSOFS,
+		ATTACKFLAGSOFS, DIAHARDOFS, REISTOPROFS,
+		KEYMGROFS, KEYMAPOFS, ACTIONBLOCKIDOFS,
+		ADDR_MAX
+	};
+
+	void initialize() {
+		/* const  wchar_t  *  const  temp[ADDR_MAX]  =  {
+		   L"ADDR_PBATTLEMGR",         L"ADDR_PNETOBJECT",       L"ADDR_COMMMODE",
+		   L"ADDR_LCHARID",            L"ADDR_RCHARID",          L"ADDR_SCENEID",
+		   L"ADDR_LCHAROFS",           L"ADDR_RCHAROFS",         L"ADDR_BTLMODEOFS",
+		   L"ADDR_LPROFOFS",           L"ADDR_RPROFOFS",         L"ADDR_NETUDPOFS",
+		   L"ADDR_ADRBEGOFS",          L"ADDR_TOADDROFS",        L"ADDR_WINCNTOFS",
+		   L"ADDR_POINTXOFS",          L"ADDR_POINTYOFS",        L"ADDR_HPOFS",
+		   L"ADDR_TIMECOUNT",          L"ADDR_SPEEDXOFS",        L"ADDR_SPEEDYOFS",
+		   L"ADDR_ACTIONIDOFS",        L"ADDR_FRAMECOUNTOFS",    L"ADDR_WEATHER",
+		   L"ADDR_DISPLAYWEATHER",     L"ADDR_WEATHERCOUNTER",   L"ADDR_ACTIVEFLAG",
+		   L"ADDR_1PKEYMAP",           L"ADDR_2PKEYMAP",         L"ADDR_GAMEDATAMGR",
+		   L"ADDR_CHAROBJHEADOFS",     L"ADDR_CHAROBJTAILOFS",   L"ADDR_OBJLISTMGR",
+		   L"ADDR_OBJPROJOFS",         L"ADDR_ATTACKAREAOFS",    L"ADDR_ATTACKAREA2OFS",
+		   L"ADDR_ATTACKAREACOUNTOFS", L"ADDR_HITAREACOUNTOFS",  L"ADDR_DIRECTIONOFS",
+		   L"ADDR_IMAGESTRUCTOFS",     L"ADDR_IMAGENUMBEROFS",   L"ADDR_HITSTOPOFS",
+		   L"ADDR_AIRDASHCOUNTOFS",    L"ADDR_REIPOWEROFS",      L"ADDR_MAXREIPOWEROFS",
+		   L"ADDR_CORRECTIONOFS",      L"ADDR_COMBORATEOFS",     L"ADDR_COMBOCOUNTOFS",
+		   L"ADDR_COMBODAMAGEOFS",     L"ADDR_COMBOLIMITOFS",    L"ADDR_CARDGAUGEOFS",
+		   L"ADDR_CARDCOUNTOFS",       L"ADDR_CARDCOUNT2OFS",    L"ADDR_HANDCARDBASEOFS",
+		   L"ADDR_SELECTCARDOFS",      L"ADDR_HANDCARDMAXOFS",   L"ADDR_SKILLLEVELMAPOFS",
+		   L"ADDR_MOTIONSTRUCTOFS",    L"ADDR_HITAREAOFS",       L"ADDR_LPROFNAME",
+		   L"ADDR_PROFNAME",           L"ADDR_ATTACKPOWEROFS",   L"ADDR_DEFENSEPOWEROFS",
+		   L"ADDR_DEFENSEDOLLOFS",     L"ADDR_ATTACKSTICKOFS",   L"ADDR_KOKUSHIOFS",
+		   L"ADDR_UTHIWAOFS",          L"ADDR_SPEEDPOWEROFS",    L"ADDR_GRIMOREOFS",
+		   L"ADDR_DROPWATEROFS",       L"ADDR_DROPWATERTIMEOFS", L"ADDR_RYUUSEIOFS",
+		   L"ADDR_TIMESTOPOFS",        L"ADDR_KANAKOOFS",        L"ADDR_SUWAKOOFS",
+		   L"ADDR_STAGENUMBER",        L"ADDR_STAGENUMBER2",     L"ADDR_BGMNUMBER",
+		   L"ADDR_SAKUYAWORLDOFS",     L"ADDR_PRIVATESQOFS",     L"ADDR_TENSEITIMEOFS",
+		   L"ADDR_TENSEINUMOFS",       L"ADDR_TENSEIONOFS",      L"ADDR_LIFERECOVERYOFS",
+		   L"ADDR_TEKETENOFS",         L"ADDR_ORRERYOFS",        L"ADDR_PHILOSOPHEROFS",
+		   L"ADDR_MUNENOFS",           L"ADDR_HIHISOUOFS",       L"ADDR_FIELDREDOFS",
+		   L"ADDR_FIELDPURPLEOFS",     L"ADDR_MPPOFS",           L"ADDR_FIELDRED2OFS",
+		   L"ADDR_HITAREAFLAGOFS",     L"ADDR_HITAREA2OFS",      L"ADDR_FRAMEFLAGSOFS",
+		   L"ADDR_ATTACKFLAGSOFS",     L"ADDR_DIAHARDOFS",       L"ADDR_REISTOPROFS",
+		   L"ADDR_KEYMGROFS",          L"ADDR_KEYMAPOFS",        L"ADDR_ACTIONBLOCKIDOFS"
+		 };
+		 ::memcpy(name_list,  temp,  sizeof(temp));
+		 reset();*/
+		 //    名称    緋想天    非想天則
+		 //84BDE0  泡姫の最大HP
+		set_default(WEATHER, 0x006E4E20, 0x008971C0);//
+		set_default(DISPLAYWEATHER, 0x006E4E24, 0x008971C4);//
+		set_default(WEATHERCOUNTER, 0x006E4E2C, 0x008971CC);//
+		set_default(STAGENUMBER, 0x00000000, 0x008971CE);//
+		set_default(TIMECOUNT, 0x006E6238, 0x008985D8);//
+		set_default(GAMEDATAMGR, 0x006E623C, 0x008985DC);//
+		set_default(CHAROBJHEADOFS, 0x3C, 0x40);      //
+		set_default(CHAROBJTAILOFS, 0x40, 0x44);      //
+		set_default(OBJLISTMGR, 0x658, 0x6F8);     //
+		set_default(OBJPROJOFS, 0x54, 0x54);      //未検証
+		set_default(PBATTLEMGR, 0x006E6244, 0x008985E4);//
+		set_default(COMMMODE, 0x006E62EC, 0x00898690);//
+		set_default(PNETOBJECT, 0x006E62FC, 0x008986A0);//
+		set_default(LPROFNAME, 0x006E6578, 0x008604A0);//未検証
+		set_default(KEYMAP1P, 0x006E6598, 0x00898940);//
+		set_default(RPROFNAME, 0x006E68B4, 0x00860C8C);//未検証
+		set_default(KEYMAP2P, 0x006E6940, 0x0089912C);//
+		set_default(STAGENUMBER2, 0x00000000, 0x00899D0C);//
+		set_default(BGMNUMBER, 0x00000000, 0x00899D0D);//
+		set_default(LCHARID, 0x006E6FF0, 0x00899D10);//
+		set_default(RCHARID, 0x006E7010, 0x00899D30);//
+		set_default(SCENEID, 0x006ECE7C, 0x008A0044);//
+		set_default(ACTIVEFLAG, 0x006ED3A8, 0x008A0F88);//
+		set_default(LCHAROFS, 0x0c, 0x0c);      //未検証
+		set_default(RCHAROFS, 0x10, 0x10);      //未検証
+		set_default(BTLMODEOFS, 0x88, 0x88);      //未検証
+		set_default(LPROFOFS, 0x04, 0x04);      //未検証
+		set_default(RPROFOFS, 0x24, 0x24);      //未検証
+		set_default(NETUDPOFS, 0x3B4, 0x3BC);     //未検証
+		set_default(ADRBEGOFS, 0x4B0, 0x10C);     //未検証
+		set_default(TOADDROFS, 0x3E0, 0x3C);      //未検証
+		set_default(POINTXOFS, 0xEC, 0xEC);      //未検証
+		set_default(POINTYOFS, 0xF0, 0xF0);      //未検証
+		set_default(SPEEDXOFS, 0xF4, 0xF4);      //未検証
+		set_default(SPEEDYOFS, 0xF8, 0xF8);      //未検証
+		set_default(DIRECTIONOFS, 0x104, 0x104);     //未検証
+		//0x11C  横倍率
+		//0x120  縦倍率
+		set_default(ACTIONIDOFS, 0x13C, 0x13C);     //未検証
+		set_default(ACTIONBLOCKIDOFS, 0x13E, 0x13E);     //未検証
+		set_default(FRAMECOUNTOFS, 0x144, 0x144);     //未検証
+		set_default(IMAGESTRUCTOFS, 0x150, 0x150);     //未検証
+		set_default(IMAGENUMBEROFS, 0x0A, 0x0A);      //未検証
+		set_default(FRAMEFLAGSOFS, 0x4C, 0x4C);      //未検証
+		set_default(ATTACKFLAGSOFS, 0x50, 0x50);      //未検証
+		set_default(MOTIONSTRUCTOFS, 0x158, 0x158);     //未検証
+		set_default(HITAREAOFS, 0x5C, 0x5C);      //未検証
+		set_default(HPOFS, 0x174, 0x184);     //未検証
+		//0x190  攻撃ヒットフラグ？　ここが0の時しか攻撃がヒットしない
+		//0x194  攻撃ヒットフラグ？　ここが1の時しか攻撃がヒットしない
+		set_default(HITSTOPOFS, 0x186, 0x196);     //未検証
+		set_default(ATTACKAREACOUNTOFS, 0x1AF, 0x1CB);     //未検証
+		set_default(HITAREACOUNTOFS, 0x1B0, 0x1CC);     //未検証
+		set_default(HITAREA2OFS, 0x1B4, 0x1D0);     //未検証
+		set_default(ATTACKAREA2OFS, 0x204, 0x220);     //未検証
+		set_default(ATTACKAREAOFS, 0x304, 0x320);     //未検証
+		set_default(HITAREAFLAGOFS, 0x318, 0x334);     //未検証
+		set_default(AIRDASHCOUNTOFS, 0x47F, 0x49B);     //未検証
+		set_default(REIPOWEROFS, 0x482, 0x49E);     //未検証
+		set_default(MAXREIPOWEROFS, 0x484, 0x4A0);     //未検証
+		set_default(REISTOPROFS, 0x486, 0x4A2);     //未検証
+		set_default(TIMESTOPOFS, 0x000, 0x4A8);     //未検証
+		set_default(CORRECTIONOFS, 0x491, 0x4AD);     //未検証
+		set_default(COMBORATEOFS, 0x494, 0x4B0);     //未検証
+		set_default(COMBOCOUNTOFS, 0x498, 0x4B4);     //未検証
+		set_default(COMBODAMAGEOFS, 0x49A, 0x4B6);     //未検証
+		set_default(COMBOLIMITOFS, 0x4A2, 0x4B8);     //未検証
+		set_default(SPEEDPOWEROFS, 0x000, 0x4D0);     //未検証
+		set_default(HIHISOUOFS, 0x4B2, 0x526);     //未検証
+		set_default(LIFERECOVERYOFS, 0x4B4, 0x528);     //未検証
+		set_default(ATTACKPOWEROFS, 0x000, 0x530);     //未検証
+		set_default(DEFENSEPOWEROFS, 0x000, 0x534);     //未検証
+		set_default(GRIMOREOFS, 0x000, 0x560);     //未検証
+		set_default(WINCNTOFS, 0x4E8, 0x573);     //未検証
+		//0x5A0なにか
+		//0x5B0デッキの残りカード枚数
+		set_default(CARDGAUGEOFS, 0x558, 0x5E4);     //未検証
+		set_default(CARDCOUNTOFS, 0x55A, 0x5E6);     //未検証
+		//0x5D7カードに関係する何か、多分手札最大値とかそんなん
+		set_default(HANDCARDBASEOFS, 0x560, 0x5EC);     //未検証
+		set_default(HANDCARDMAXOFS, 0x564, 0x5F0);     //未検証
+		set_default(SELECTCARDOFS, 0x568, 0x5F4);     //未検証
+		set_default(CARDCOUNT2OFS, 0x56C, 0x5F8);     //未検証
+		//set_default(SKILLMAPADDROFS,  0x604,      0x694);     //未検証  未使用)
+		set_default(SKILLLEVELMAPOFS, 0x624, 0x6C4);     //未検証
+		set_default(KEYMGROFS, 0x000, 0x750);     //未検証
+		set_default(KEYMAPOFS, 0x000, 0x38);      //未検証
+		//0x7BC(カードに関係する何か)
+		set_default(UTHIWAOFS, 0x000, 0x834);     //未検証
+		set_default(DEFENSEDOLLOFS, 0x000, 0x840);     //未検証
+		set_default(ATTACKSTICKOFS, 0x000, 0x844);     //未検証
+		set_default(RYUUSEIOFS, 0x000, 0x84E);     //未検証
+		set_default(DROPWATEROFS, 0x000, 0x850);     //未検証
+		set_default(DROPWATERTIMEOFS, 0x000, 0x852);     //未検証
+		set_default(SAKUYAWORLDOFS, 0x7A4, 0x890);     //未検証
+		set_default(PHILOSOPHEROFS, 0x7A4, 0x890);     //未検証
+		set_default(PRIVATESQOFS, 0x7A6, 0x892);     //未検証
+		set_default(ORRERYOFS, 0x7A6, 0x892);     //未検証
+		set_default(MPPOFS, 0x7A6, 0x892);     //未検証
+		set_default(FIELDREDOFS, 0x7AC, 0x898);     //未検証
+		set_default(FIELDPURPLEOFS, 0x7AE, 0x89A);     //未検証
+		set_default(KANAKOOFS, 0x000, 0x89C);     //未検証
+		set_default(SUWAKOOFS, 0x000, 0x8A0);     //未検証
+		set_default(DIAHARDOFS, 0x000, 0x8A0);     //未検証
+		set_default(KOKUSHIOFS, 0x000, 0x8A0);     //未検証
+		set_default(FIELDRED2OFS, 0x000, 0x8B2);     //未検証
+		set_default(TENSEITIMEOFS, 0x7C6, 0x8B4);     //未検証
+		set_default(TENSEINUMOFS, 0x7C8, 0x8B6);     //未検証
+		set_default(TENSEIONOFS, 0x7CA, 0x8B8);     //未検証
+		set_default(TEKETENOFS, 0x7EA, 0x8D6);     //未検証
+		set_default(MUNENOFS, 0x838, 0x924);     //
+	}
+	//const wchar_t *name_list[ADDR_MAX];
+	unsigned int list[ADDR_MAX];
+
+	const unsigned int &operator [](unsigned int n) {
+		/*if(list[n] == 0)  {
+		  list[n] = g_ini.GetInt(name_list[n]);
+		}*/
+		return list[n];
+	}
+
+	/*void reset() {
+	  ::memset(list, 0, sizeof(list));
+	}*/
+private:
+	void set_default(unsigned int n, unsigned int swr_addr, unsigned int swrs_addr) {
+		/*wchar_t addr[256], name[256];
+		::_snwprintf(name, _countof(name), L"SWR_%s", name_list[n]);
+		::_snwprintf(addr, _countof(addr), L"0x%x", swr_addr);
+		g_ini.Add(L"default", name, addr);
+		::_snwprintf(name, _countof(name), L"SWRS_%s", name_list[n]);
+		::_snwprintf(addr, _countof(addr), L"0x%x", swrs_addr);
+		g_ini.Add(L"default", name, addr);*/
+		list[n] = swrs_addr;
+	}
+};
+
+//extern address addr;
+
+#endif // address_h__
