+++ LabTool/src/stdafx.h
@@ -1,0 +1,34 @@
+// stdafx.h : include file for standard system include files,
+// or project specific include files that are used frequently, but
+// are changed infrequently
+//
+
+#pragma once
+#ifndef stdafx_h__
+#define stdafx_h__
+
+#include "targetver.h"
+
+#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
+// Windows Header Files
+#include <windows.h>
+
+//Standard C++ headers
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <iostream>
+
+//Standard headers
+#include <array>
+#include <d3d9.h>
+#include <mmsystem.h>
+#include <shlwapi.h>
+#include <sstream>
+#include <string>
+#include <windows.h>
+
+
+// reference additional headers your program requires here
+
+#endif // stdafx_h__
