+++ LabTool/src/targetver.h
@@ -1,0 +1,11 @@
+#pragma once
+#ifndef targetver_h__
+#define targetver_h__
+// Including SDKDDKVer.h defines the highest available Windows platform.
+
+// If you wish to build your application for a previous Windows platform, include WinSDKVer.h and
+// set the _WIN32_WINNT macro to the platform you wish to support before including SDKDDKVer.h.
+
+#include <SDKDDKVer.h>
+
+#endif // targetver_h__
